Starting from interfaces.
