+++ src/main/java/com/revature/services/AuthServiceImpl.java
@@ -1,0 +1,24 @@
+package com.revature.services;
+
+import com.revature.models.User;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+public class AuthServiceImpl {
+
+    private final UserServiceImpl userService;
+
+    public AuthServiceImpl(UserServiceImpl userService) {
+        this.userService = userService;
+    }
+
+    public Optional<User> findByCredentials(String email, String password) {
+        return userService.findByCredentials(email, password);
+    }
+
+    public User register(User user) {
+        return userService.save(user);
+    }
+}
