+++ src/main/java/com/revature/services/UserService.java
@@ -1,25 +1,11 @@
 package com.revature.services;
 
 import com.revature.models.User;
-import com.revature.repositories.UserRepository;
-import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
-@Service
-public class UserService {
+public interface UserService {
+    public Optional<User> findByCredentials(String email, String password);
 
-    private final UserRepository userRepository;
-
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    public Optional<User> findByCredentials(String email, String password) {
-        return userRepository.findByEmailAndPassword(email, password);
-    }
-
-    public User save(User user) {
-        return userRepository.save(user);
-    }
+    public User save(User user);
 }
